tableau_sql_agent/pyproject.toml
[project]
name = "tableau-sql-agent"
version = "0.1.0"
description = "Tableau to PostgreSQL migration accelerator using LangGraph and SQL Agent"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-community>=0.3.0",
    "google-cloud-aiplatform>=1.70.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "xmltodict>=0.13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.setuptools.packages.find]
include = ["database*", "tools*", "agents*"]
